+++ src/index.js
@@ -8,7 +8,7 @@ export default {
   init: function(options) {
     YFClient.init(options)
   },
-
+  ping: YFClient.ping,
   Func: Func,
   Query: Query,
   Object: Obj,